all done

+++ services/server/menus/categoryMenus.js
@@ -1,5 +1,6 @@
 import { errorHandler } from "../errorhandler";
 import { ObjectId } from "mongodb";
+import { userVarification } from "../user/user";
 
 export async function getCategoryMenus(req, res, category) {
   try {
@@ -12,11 +13,16 @@ export async function getCategoryMenus(req, res, category) {
 
 export async function postCategoryMenus(req, res, categoryMenus) {
   try {
+    if (!req.body.userId) throw { message: "user unathenticated!" };
+    const { varify } = await userVarification(req.body.userId);
+    if (!varify) throw { message: "user unathenticated!" };
+
     const isExist = await categoryMenus.findOne({ name: req.body.name });
     if (isExist) {
       res.status(409).send({ message: "Already added the category" });
       return;
     }
+    delete req.body.userId;
     const result = await categoryMenus.insertOne(req.body);
     if (result.insertedId) {
       res.status(200).send({
@@ -32,6 +38,10 @@ export async function postCategoryMenus(req, res, categoryMenus) {
 
 export async function updateCategoryMenus(req, res, categoryMenus) {
   try {
+    if (!req.body.userId) throw { message: "user unathenticated!" };
+    const { varify } = await userVarification(req.body.userId);
+    if (!varify) throw { message: "user unathenticated!" };
+    delete req.body.userId;
     const result = await categoryMenus.updateOne(
       {
         _id: ObjectId(req.body.categoryId),
@@ -53,6 +63,11 @@ export async function updateCategoryMenus(req, res, categoryMenus) {
 
 export async function deleteCategoryMenu(req, res, category) {
   try {
+    if (!req.body.userId) throw { message: "user unathenticated!" };
+    const { varify } = await userVarification(req.body.userId);
+    if (!varify) throw { message: "user unathenticated!" };
+    delete req.body.userId;
+
     const result = await category.deleteOne({ _id: ObjectId(req.body.id) });
     if (result.deletedCount > 0) {
       res.status(200).send({
@@ -62,6 +77,6 @@ export async function deleteCategoryMenu(req, res, category) {
       res.status(424).send({ message: "Unable to delete, Try again." });
     }
   } catch (error) {
-    errorHandler(res);
+    errorHandler(res, { msg: error.message, status: error.status });
   }
 }
